refactor: split nixgraph parsing and rendering helpers
Extract nix-store graph parsing, store-path resolution,
and graph traversal and rendering into focused nixgraph
modules. Keep nixgraph.graph as the public orchestration
surface so the CLI and exception tests continue to patch
and exercise the same entrypoints.

Add fast unit coverage for parsed dependency edges and
CSV-mode graph traversal while preserving existing CLI
behavior.
